requirements.txt
certbot==1.27.0

# cetbot plugins for DNS providers
certbot-dns-cloudflare==1.27.0
certbot-dns-cloudxns==1.27.0
certbot-dns-digitalocean==1.27.0
certbot-dns-dnsimple==1.27.0
certbot-dns-dnsmadeeasy==1.27.0
certbot-dns-gehirn==1.27.0
certbot-dns-google==1.27.0
certbot-dns-linode==1.27.0
certbot-dns-luadns==1.27.0
certbot-dns-nsone==1.27.0
certbot-dns-ovh==1.27.0
certbot-dns-rfc2136==1.27.0
certbot-dns-route53==1.27.0
certbot-dns-sakuracloud==1.27.0
certbot-dns-godaddy==0.2.2

# fixing versions of dependent packages
cloudflare>=2.3.1

# flask with fixed dependencies
flask==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
click==8.1.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
zipp==3.8.0
urllib3==1.26.9
requests==2.27.1
certifi>=2021.10.8
CherryPy==18.6.1
tornado==6.1
chardet==4.0.0
cheroot==8.6.0
marshmallow==3.15.0
marshmallow-jsonschema==0.13.0

# google secrets manager client
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.3.0
pytz>=2022.1